IGNITE-14917 <compilation error>
