+++ modules/configuration-annotation-processor/pom.xml
@@ -105,7 +105,7 @@
                         <configuration>
                             <compilerArgument>-proc:none</compilerArgument>
                             <includes>
-                                <include>org/apache/ignite/configuration/processor/internal/*</include>
+                                <include>org/apache/ignite/internal/configuration/processor/*</include>
                             </includes>
                         </configuration>
                     </execution>
