+++ modules/configuration-annotation-processor/src/test/java/org/apache/ignite/internal/configuration/sample/CacheConfigurationSchema.java
@@ -15,22 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.configuration.sample;
+package org.apache.ignite.internal.configuration.sample;
 
-import org.apache.ignite.configuration.annotation.ConfigValue;
-import org.apache.ignite.configuration.annotation.ConfigurationRoot;
-import org.apache.ignite.configuration.annotation.NamedConfigValue;
+import org.apache.ignite.configuration.annotation.Config;
+import org.apache.ignite.configuration.annotation.Value;
+import org.apache.ignite.configuration.validation.Min;
 
 /**
- * Test cluster wide configuration schema.
+ * Test cache configuration schema.
  */
-@ConfigurationRoot(rootName = "cluster")
-public class ClusterWideConfigurationSchema {
-    /** Cache. */
-    @NamedConfigValue
-    public CacheConfigurationSchema cacheConfig;
-
-    /** Baseline. */
-    @ConfigValue
-    public BaselineConfigurationSchema baseline;
+@Config
+public class CacheConfigurationSchema {
+    /** Size. */
+    @Value
+    @Min(1)
+    public int size;
 }
