+++ modules/configuration-annotation-processor/src/test/java/org/apache/ignite/internal/configuration/sample/ClusterWideConfigurationSchema.java
@@ -15,23 +15,22 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.configuration.sample;
+package org.apache.ignite.internal.configuration.sample;
 
-import org.apache.ignite.configuration.annotation.Config;
-import org.apache.ignite.configuration.annotation.Value;
-import org.apache.ignite.configuration.validation.Min;
+import org.apache.ignite.configuration.annotation.ConfigValue;
+import org.apache.ignite.configuration.annotation.ConfigurationRoot;
+import org.apache.ignite.configuration.annotation.NamedConfigValue;
 
 /**
- * Test auto adjust configuration schema.
+ * Test cluster wide configuration schema.
  */
-@Config
-public class AutoAdjustConfigurationSchema {
-    /** Timeout. */
-    @Value(hasDefault = true)
-    @Min(0)
-    public long timeout = 0L;
+@ConfigurationRoot(rootName = "cluster")
+public class ClusterWideConfigurationSchema {
+    /** Cache. */
+    @NamedConfigValue
+    public CacheConfigurationSchema cacheConfig;
 
-    /** Enabled. */
-    @Value(hasDefault = true)
-    public boolean enabled = true;
+    /** Baseline. */
+    @ConfigValue
+    public BaselineConfigurationSchema baseline;
 }
