+++ modules/configuration-annotation-processor/src/test/java/org/apache/ignite/internal/configuration/sample/NodeConfigurationSchema.java
@@ -15,19 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.configuration.sample;
+package org.apache.ignite.internal.configuration.sample;
 
 import org.apache.ignite.configuration.annotation.Config;
 import org.apache.ignite.configuration.annotation.Value;
-import org.apache.ignite.configuration.validation.Min;
+import org.apache.ignite.configuration.validation.Immutable;
 
 /**
- * Test cache configuration schema.
+ * Test node configuration schema.
  */
 @Config
-public class CacheConfigurationSchema {
-    /** Size. */
+public class NodeConfigurationSchema {
+    /** Consistent id. */
     @Value
-    @Min(1)
-    public int size;
+    @Immutable
+    public String consistentId;
+
+    /** Port. */
+    @Value
+    public int port;
+
+    /** Auto adjust enabled. */
+    @Value(hasDefault = true)
+    public boolean autoAdjustEnabled = true;
 }
